Module governance_os_node::service

Structs

NativeExecutor
A generic CodeExecutor implementation that uses a delegate to determine wasm code equivalence and dispatch to native code when possible, falling back on WasmExecutor when not.

Functions

new_full
Builds a new service for a full client.

new_light
Builds a new service for a light client.

new_partial

remote_keystore

Type Definitions

FullBackend

FullClient

FullSelectChain